Replaced old Header and removed Body text
