--- application/frontend/src/pages/Search/components/BodyText.scss
@@ -1,14 +1,0 @@
-.index-text {
-    margin-top: 2rem;
-    padding: 15px;
-    width: 70%;
-    background: white;
-}
-
-// mobile
-@media (min-width: 0px) and (max-width: 599px) {
-    .index-text {
-        width: 90%;
-    }
-}
-
