+++ application/frontend/src/pages/Search/components/SearchBar.scss
@@ -1,9 +1,46 @@
-#SearchBar {
-  margin: auto;
-  padding: 0;
-}
+.navbar__search {
+  display: none;
+  align-items: center;
 
-#SearchButton {
-  margin: auto;
-  padding: 0;
+  @media (min-width: 1024px) {
+    display: flex;
+  }
+
+  form {
+    position: relative;
+  }
+
+  .search-icon {
+    position: absolute;
+    left: 0.75rem;
+    top: 50%;
+    transform: translateY(-50%);
+    color: var(--muted-foreground);
+    height: 1rem;
+    width: 1rem;
+  }
+
+  input {
+    padding: 0.5rem 1rem 0.5rem 2.5rem;
+    width: 16rem;
+    background-color: rgba(var(--card-rgb), 0.5);
+    backdrop-filter: blur(4px);
+    border: 1px solid rgba(var(--border-rgb), 0.5);
+    border-radius: var(--radius);
+    color: var(--foreground);
+    transition: all 0.2s ease;
+
+    &:focus {
+      outline: none;
+      border-color: transparent; // remove default
+      box-shadow: 0 0 0 2px #ffffff, 0 0 0 4px #3b82f6;
+      /* 1st layer white, 2nd layer blue */
+    }
+  }
+
+  .search-error {
+    margin-top: 0.25rem;
+    font-size: 0.75rem;
+    color: #f87171; // red-400
+  }
 }
