--- application/frontend/src/pages/Search/components/navbar.scss
@@ -1,298 +1,0 @@
-/* navbar.scss */
-
-//==========================================================================
-// 1. NAVBAR CONTAINER & LAYOUT
-//==========================================================================
-
-.navbar {
-  // Positioning and background
-  position: sticky;
-  top: 0;
-  z-index: 50;
-  background-color: rgba(var(--background-rgb), 0.8);
-  backdrop-filter: blur(4px);
-  border-bottom: 1px solid rgba(var(--border-rgb), 0.4);
-
-  &__container {
-    max-width: 112rem; // max-w-7xl
-    margin: 0 auto;
-    padding: 0 1rem;
-    @media (min-width: 640px) {
-      padding: 0 1.5rem;
-    }
-    @media (min-width: 1024px) {
-      padding: 0 2rem;
-    }
-  }
-
-  &__content {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 4rem; // h-16
-  }
-}
-
-//==========================================================================
-// 2. NAVBAR SECTIONS (LOGO, LINKS, AUTH)
-//==========================================================================
-
-.navbar__logo {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  img {
-    height: 10rem;
-    width: 10rem;
-  }
-}
-
-.navbar__desktop-links {
-  display: none; // Hidden by default
-  align-items: center;
-  gap: 2rem;
-
-  @media (min-width: 768px) {
-    // md:flex
-    display: flex;
-  }
-
-  .nav-link {
-    color: var(--muted-foreground);
-    text-decoration: none;
-    transition: color 0.2s;
-    &:hover {
-      color: var(--foreground);
-    }
-  }
-}
-
-.navbar__search {
-  display: none; // hidden by default
-  align-items: center;
-
-  @media (min-width: 1024px) {
-    // lg:flex
-    display: flex;
-  }
-
-  form {
-    position: relative;
-  }
-
-  .search-icon {
-    position: absolute;
-    left: 0.75rem;
-    top: 50%;
-    transform: translateY(-50%);
-    color: var(--muted-foreground);
-    height: 1rem;
-    width: 1rem;
-  }
-
-  input {
-    padding: 0.5rem 1rem 0.5rem 2.5rem; // py-2 pr-4 pl-10
-    width: 16rem; // w-64
-    background-color: rgba(var(--card-rgb), 0.5);
-    backdrop-filter: blur(4px);
-    border: 1px solid rgba(var(--border-rgb), 0.5);
-    border-radius: var(--radius);
-    color: var(--foreground);
-    transition: border-color 0.2s;
-
-    &:focus {
-      outline: none;
-      border-color: #60a5fa; // focus:border-blue-400
-    }
-  }
-}
-
-.navbar__actions {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.navbar__desktop-auth {
-  display: none; // Hidden by default
-  align-items: center;
-  gap: 1rem;
-
-  @media (min-width: 768px) {
-    // md:flex
-    display: flex;
-  }
-
-  .user-info {
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-    .user-details {
-      display: flex;
-      align-items: center;
-      gap: 0.5rem;
-      .icon {
-        height: 1rem;
-        width: 1rem;
-      }
-      span {
-        font-size: 0.875rem;
-        color: var(--muted-foreground);
-      }
-    }
-  }
-
-  // Button styles
-  .btn {
-    padding: 0.5rem 1rem;
-    border-radius: var(--radius);
-    border: none;
-    cursor: pointer;
-    transition: all 0.2s;
-    &--ghost {
-      background: transparent;
-      color: var(--muted-foreground);
-      &:hover {
-        color: var(--foreground);
-      }
-    }
-    &--primary {
-      background-color: white;
-      color: black;
-      &:hover {
-        background-color: #e5e7eb; // gray-200
-      }
-    }
-  }
-}
-
-//==========================================================================
-// 3. MOBILE MENU
-//==========================================================================
-
-.navbar__mobile-menu-toggle {
-  background: transparent;
-  border: none;
-  cursor: pointer;
-  padding: 0.5rem;
-  border-radius: var(--radius);
-
-  .icon {
-    height: 1.25rem;
-    width: 1.25rem;
-    color: var(--foreground);
-  }
-
-  @media (min-width: 768px) {
-    // md:hidden
-    display: none;
-  }
-}
-
-.navbar__mobile-menu {
-  position: fixed;
-  top: 0;
-  right: 0;
-  width: 300px;
-  height: 100%;
-  background-color: var(--background);
-  border-left: 1px solid var(--border);
-  padding: 1rem;
-  transform: translateX(100%);
-  transition: transform 0.3s ease-in-out;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-
-  &.is-open {
-    transform: translateX(0);
-  }
-  @media (min-width: 400px) {
-    // sm:w-[400px]
-    width: 400px;
-  }
-
-  .mobile-search-form {
-    position: relative;
-    .search-icon {
-      position: absolute;
-      left: 0.75rem;
-      top: 50%;
-      transform: translateY(-50%);
-      color: var(--muted-foreground);
-      height: 1rem;
-      width: 1rem;
-    }
-    input {
-      width: 100%;
-      padding: 0.5rem 1rem 0.5rem 2.5rem;
-      background-color: var(--input);
-      border: 1px solid var(--border);
-      border-radius: var(--radius);
-      color: var(--foreground);
-    }
-  }
-
-  .mobile-nav-links {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    .nav-link {
-      padding: 0.5rem 1rem;
-      border-radius: var(--radius);
-      color: var(--muted-foreground);
-      text-decoration: none;
-      transition: all 0.2s;
-      &:hover {
-        color: var(--foreground);
-        background-color: var(--muted);
-      }
-    }
-  }
-
-  .mobile-auth {
-    border-top: 1px solid var(--border);
-    padding-top: 1rem;
-
-    .user-info,
-    .auth-buttons {
-      display: flex;
-      flex-direction: column;
-      gap: 0.75rem;
-    }
-
-    .user-details {
-      display: flex;
-      align-items: center;
-      gap: 0.5rem;
-      padding: 0 1rem;
-      .icon {
-        height: 1rem;
-        width: 1rem;
-      }
-      span {
-        font-size: 0.875rem;
-        color: var(--muted-foreground);
-      }
-    }
-
-    .btn {
-      width: 100%;
-      text-align: left;
-      justify-content: flex-start;
-    }
-  }
-}
-
-// Overlay for when mobile menu is open
-.navbar__overlay {
-  position: fixed;
-  inset: 0;
-  z-index: 40; // Below navbar mobile menu
-  background-color: rgba(0, 0, 0, 0.5);
-  display: none;
-
-  &.is-open {
-    display: block;
-  }
-}
