+++ application/frontend/src/scaffolding/Header/header.scss
@@ -1,105 +1,259 @@
-.header__nav-bar {
-  padding: 0 10px;
+//==========================================================================
+// 1. NAVBAR CONTAINER & LAYOUT
+//==========================================================================
 
-  &.ui.secondary.menu {
-    margin: 0;
-    padding: 0 30px;
-    background-color: #232e4c;
+.navbar {
+  // Positioning and background
+  position: sticky;
+  top: 0;
+  z-index: 50;
+  background-color: rgba(var(--background-rgb), 0.8);
+  backdrop-filter: blur(4px);
+  border-bottom: 1px solid rgba(255, 255, 255, 0.2);
 
-    .header__nav-bar-logo {
-      padding: 20px 0;
-      margin-right: 20px;
+  &__container {
+    max-width: 112rem; // max-w-7xl
+    margin: 0 auto;
+    padding: 0 1rem;
+    @media (min-width: 640px) {
+      padding: 0 1.5rem;
+    }
+    @media (min-width: 1024px) {
+      padding: 0 2rem;
+    }
+  }
+
+  &__content {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 4rem; // h-16
+  }
+}
+
+//==========================================================================
+// 2. NAVBAR SECTIONS (LOGO, LINKS, AUTH)
+//==========================================================================
+
+.navbar__logo {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  margin-left: 4rem;
+  img {
+    height: 10rem;
+    width: 10rem;
+  }
+}
+
+.navbar__desktop-links {
+  display: none; // Hidden by default
+  align-items: center;
+  gap: 2rem;
+
+  @media (min-width: 768px) {
+    // md:flex
+    display: flex;
+  }
+
+  .nav-link {
+    color: var(--muted-foreground);
+    text-decoration: none;
+    font-size: 1.2rem; // text-lg
+
+    transition: color 0.2s;
+    &:hover {
+      color: var(--foreground);
+    }
+  }
+}
+
+.navbar__actions {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.navbar__desktop-auth {
+  display: none; // Hidden by default
+  align-items: center;
+  gap: 1rem;
+
+  @media (min-width: 768px) {
+    // md:flex
+    display: flex;
+  }
+
+  .user-info {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    .user-details {
       display: flex;
       align-items: center;
-      max-width: 140px;
-
-      img {
-        max-width: 100px;
-        max-height: 30px;
+      gap: 0.5rem;
+      .icon {
+        height: 1rem;
+        width: 1rem;
       }
-
-      h1 {
-        display: inline;
-        margin: 0 10px;
-        color: white;
-        font-size: 20px;
+      span {
+        font-size: 0.875rem;
+        color: var(--muted-foreground);
       }
     }
-    .item {
-      color: white;
-      font-size: 14px;
-      .ui.form {
-        .fields {
-          margin-bottom: 0px !important;
-        }
+  }
+
+  // Button styles
+  .btn {
+    padding: 0.5rem 1rem;
+    border-radius: var(--radius);
+    border: none;
+    cursor: pointer;
+    transition: all 0.2s;
+    &--ghost {
+      background: transparent;
+      color: var(--muted-foreground);
+      &:hover {
+        color: var(--foreground);
       }
     }
-
-    .header__nav-bar__link {
-      padding-top: 10px;
-      padding-bottom: 10px;
-
-      .item {
-        width: 100%;
-        height: 100%;
-        margin: 0;
-        text-align: center;
-      }
-      &:hover:not(.header__nav-bar__link--active) .item {
-        text-decoration: underline;
-      }
-      &--active {
-        transition: all 0.2s ease;
-
-        .item {
-          font-weight: bold;
-          transition: all 0.2s ease;
-          border-radius: 0 !important;
-          background-color: transparent !important;
-        }
+    &--primary {
+      background-color: white;
+      color: black;
+      &:hover {
+        background-color: #e5e7eb; // gray-200
       }
     }
   }
 }
 
-nav + * {
-  padding-top: 30px;
-}
+//==========================================================================
+// 3. MOBILE MENU
+//==========================================================================
 
-// mobile
-@media (min-width: 0px) and (max-width: 799px) {
-  .header__nav-bar {
-    flex-wrap: wrap;
-    .right.menu {
-      display: block;
-    }
-    &__link {
-      padding-top: 0;
-      padding-bottom: 0;
-    }
+.navbar__mobile-menu-toggle {
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  padding: 0.5rem;
+  border-radius: var(--radius);
+
+  .icon {
+    height: 1.25rem;
+    width: 1.25rem;
+    color: var(--foreground);
+  }
+
+  @media (min-width: 768px) {
+    // md:hidden
+    display: none;
   }
 }
 
-@media (min-width: 0px) and (max-width: 770px) {
-  .header__nav-bar {
-    justify-content: center;
+.navbar__mobile-menu {
+  position: fixed;
+  top: 0;
+  right: 0;
+  width: 300px;
+  height: 100%;
+  background-color: var(--background);
+  border-left: 1px solid var(--border);
+  padding: 1rem;
+  transform: translateX(100%);
+  transition: transform 0.3s ease-in-out;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
 
-    &.ui.secondary.menu {
-      padding: 0 0 10px;
-      flex-direction: column;
-      align-items: center;
+  &.is-open {
+    transform: translateX(0);
+  }
+  @media (min-width: 400px) {
+    // sm:w-[400px]
+    width: 400px;
+  }
 
-      .header__nav-bar-logo {
-        margin-right: 0;
-        padding: 20px 0 0;
-      }
+  .mobile-search-form {
+    position: relative;
+    .search-icon {
+      position: absolute;
+      left: 0.75rem;
+      top: 50%;
+      transform: translateY(-50%);
+      color: var(--muted-foreground);
+      height: 1rem;
+      width: 1rem;
     }
-
-    &.ui.menu:not(.vertical) .right.menu {
-      display: flex;
-      justify-content: center;
+    input {
       width: 100%;
-      margin-left: unset !important;
+      padding: 0.5rem 1rem 0.5rem 2.5rem;
+      background-color: var(--input);
+      border: 1px solid var(--border);
+      border-radius: var(--radius);
+      color: var(--foreground);
     }
+  }
+
+  .mobile-nav-links {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    .nav-link {
+      padding: 0.5rem 1rem;
+      border-radius: var(--radius);
+      color: var(--muted-foreground);
+      text-decoration: none;
+      transition: all 0.2s;
+      &:hover {
+        color: var(--foreground);
+        background-color: var(--muted);
+      }
+    }
+  }
+
+  .mobile-auth {
+    border-top: 1px solid var(--border);
+    padding-top: 1rem;
+
+    .user-info,
+    .auth-buttons {
+      display: flex;
+      flex-direction: column;
+      gap: 0.75rem;
+    }
+
+    .user-details {
+      display: flex;
+      align-items: center;
+      gap: 0.5rem;
+      padding: 0 1rem;
+      .icon {
+        height: 1rem;
+        width: 1rem;
+      }
+      span {
+        font-size: 0.875rem;
+        color: var(--muted-foreground);
+      }
+    }
+
+    .btn {
+      width: 100%;
+      text-align: left;
+      justify-content: flex-start;
+    }
+  }
+}
+
+// Overlay for when mobile menu is open
+.navbar__overlay {
+  position: fixed;
+  inset: 0;
+  z-index: 40; // Below navbar mobile menu
+  background-color: rgba(0, 0, 0, 0.5);
+  display: none;
+
+  &.is-open {
+    display: block;
   }
 }
